Move MimeUtils to covia.utils; remove empty JVMTest stub
MimeUtils is a generic mime-type helper with no adapter-specific logic;
it belongs in a utilities package, not under covia.adapter. Moves it so
DLFSAdapter and FileAdapter can use it without an implicit "neighbour"
relationship to the adapter framework.

JVMTest.java was a one-byte stub (whitespace only) — leftover from an
earlier scaffold. Deleting.
