+++ venue/src/main/java/covia/adapter/FileAdapter.java
@@ -32,6 +32,7 @@ import convex.core.data.prim.CVMLong;
 import convex.core.lang.RT;
 import convex.core.util.JSON;
 import covia.grid.Asset;
+import covia.utils.MimeUtils;
 import covia.venue.Config;
 import covia.venue.Engine;
 import covia.venue.RequestContext;
