--- venue/src/main/java/covia/adapter/JVMTest.java
@@ -1,1 +1,0 @@
-
